+++ src/index.js
@@ -1,6 +1,8 @@
 "use strict";
 
 const _ = require("lodash");
+
+const AsyncApi = require("./AsyncApi");
 const SyncApi = require("./SyncApi");
 
 function argumentsToMultiLanguageObj(argsObj) {
@@ -44,18 +46,26 @@ function argumentsToMultiLanguageObj(argsObj) {
 
 /**
  *
- * @param baseUrl {string}
- * @param options {object}
- * @returns {{api: SyncApi, Table: Table, Tables: Tables, TableBuilder: TableBuilder, ColumnBuilder: ColumnBuilder,
- *   ConstraintBuilder: ConstraintBuilder}}
+ *  @param baseUrl {string}
+ *  @param options {object}
+ *  @returns {{
+ *    api: SyncApi,
+ *    asyncApi: AsyncApi,
+ *    Table: Table,
+ *    Tables: Tables,
+ *    TableBuilder: TableBuilder,
+ *    ColumnBuilder: ColumnBuilder,
+ *    ConstraintBuilder: ConstraintBuilder
+ *  }}
  */
 function grudStructorizer(baseUrl, options) {
 
-  const tableaux = new SyncApi(baseUrl, options);
+  const syncApi = new SyncApi(baseUrl, options);
+  const asyncApi = new AsyncApi(baseUrl, options);
 
   const StaticHelpers = {
     getLanguages: () => {
-      return tableaux.doCall("GET", "/system/settings/langtags").value;
+      return syncApi.doCall("GET", "/system/settings/langtags").value;
     },
 
     checkKindForLanguageConversion: (kind) => {
@@ -91,7 +101,7 @@ function grudStructorizer(baseUrl, options) {
      * @returns {Tables}
      */
     fetch() {
-      Object.assign(this, tableaux.doCall("GET", "/tables"));
+      Object.assign(this, syncApi.doCall("GET", "/tables"));
       return this;
     }
 
@@ -147,7 +157,7 @@ function grudStructorizer(baseUrl, options) {
      * @returns {Table}
      */
     fetch(includeRows = false) {
-      Object.assign(this, tableaux.fetchTable(this.tableId, includeRows));
+      Object.assign(this, syncApi.fetchTable(this.tableId, includeRows));
       return this;
     }
 
@@ -244,7 +254,7 @@ function grudStructorizer(baseUrl, options) {
         });
       });
 
-      const newColumns = tableaux.createColumns(this.tableId, columnObjArray);
+      const newColumns = syncApi.createColumns(this.tableId, columnObjArray);
 
       newColumns.forEach(function (newColumn) {
         self.columns.push(newColumn);
@@ -264,7 +274,7 @@ function grudStructorizer(baseUrl, options) {
         throw new Error("No column with this name or ID found '" + nameOrId + "'");
       }
 
-      const response = tableaux.doCall("DELETE", "/tables/" + this.tableId + "/columns/" + column.id);
+      const response = syncApi.doCall("DELETE", "/tables/" + this.tableId + "/columns/" + column.id);
 
       if (response) {
         _.remove(this.columns, function (c) {
@@ -291,7 +301,7 @@ function grudStructorizer(baseUrl, options) {
         }
       });
 
-      const newColumn = tableaux.createColumn(this.tableId, columnBuilder.build());
+      const newColumn = syncApi.createColumn(this.tableId, columnBuilder.build());
 
       this.columns.push(newColumn);
 
@@ -371,11 +381,11 @@ function grudStructorizer(baseUrl, options) {
       // generate IDs based on rowValues length
       const columnIds = columns || _.range(1, firstRowValues.length + 1);
 
-      return tableaux.createRows(this.tableId, columnIds, rows);
+      return syncApi.createRows(this.tableId, columnIds, rows);
     }
 
     changeColumn(columnId, changeObj) {
-      return tableaux.doCall("POST", "/tables/" + this.tableId + "/columns/" + columnId, changeObj);
+      return syncApi.doCall("POST", "/tables/" + this.tableId + "/columns/" + columnId, changeObj);
     }
 
     getColumn(columnName) {
@@ -436,16 +446,16 @@ function grudStructorizer(baseUrl, options) {
 
         const newValue = mapValueIntoLanguage(value, pickLanguage || defaultLanguage);
 
-        tableaux.doCall("PATCH", url, newValue);
+        syncApi.doCall("PATCH", url, newValue);
 
-        tableaux.doCall("POST", `${url}/annotations`, {
+        syncApi.doCall("POST", `${url}/annotations`, {
           langtags: languages,
           type: "flag",
           value: "needs_translation"
         });
       });
 
-      tableaux.doCall("DELETE", "/tables/" + this.tableId + "/columns/" + column.id);
+      syncApi.doCall("DELETE", "/tables/" + this.tableId + "/columns/" + column.id);
     };
 
     /**
@@ -488,19 +498,19 @@ function grudStructorizer(baseUrl, options) {
           return;
         }
 
-        tableaux.doCall("PATCH", url, { value: newValue });
+        syncApi.doCall("PATCH", url, { value: newValue });
 
         if (_.includes(annotations, columnIndex)) {
           // there schould be not more than one translation flag per cell
           const langAnnotation = _.head(_.filter(annotations[columnIndex], { "value": "needs_translation" }));
 
           if (langAnnotation) {
-            tableaux.doCall("DELETE", `${url}/annotations/${langAnnotation.uuid}`);
+            syncApi.doCall("DELETE", `${url}/annotations/${langAnnotation.uuid}`);
           }
         }
       });
 
-      tableaux.doCall("DELETE", "/tables/" + this.tableId + "/columns/" + column.id);
+      syncApi.doCall("DELETE", "/tables/" + this.tableId + "/columns/" + column.id);
     }
   }
 
@@ -566,7 +576,7 @@ function grudStructorizer(baseUrl, options) {
      * @returns {Table}
      */
     create() {
-      const tableId = tableaux.createTable(this.name, this._hidden, this._displayName, this.type, this._groupId).id;
+      const tableId = syncApi.createTable(this.name, this._hidden, this._displayName, this.type, this._groupId).id;
 
       return new Table(tableId, this.name);
     }
@@ -954,7 +964,8 @@ function grudStructorizer(baseUrl, options) {
   }
 
   return {
-    api: tableaux,
+    api: syncApi,
+    asyncApi: asyncApi,
 
     Table: Table,
     Tables: Tables,